requirements.txt
flask
ultralytics
opencv-python
pillow
numpy
torch